+++ R/as-ypr-population.R
@@ -11,5 +11,6 @@ as_ypr_population <- function(x, ...) UseMethod("as_ypr_population")
 #' @describeIn as_ypr_population Coerce data frame to ypr_population
 #' @export
 as_ypr_population.data.frame <- function(x, ...) {
+  chk_unused(...)
   do.call("ypr_population", x)
 }