+++ R/namespace.R
@@ -1,4 +1,5 @@
 #' @import ggplot2 chk yesno
+#' @importFrom purrr map map2 transpose
 #' @importFrom stats setNames
 #' @importFrom graphics plot
 NULL